Renomage des dossier pouréviter les conflits

+++ ssd/agent/entry.py
@@ -5,12 +5,15 @@ node = "https://uneurl:8000/"
 backup_dir = "."
 fallback = ["https://url1:8000", "https://url2:8000/"]
 """
+import datetime
 import os
 import time
 
 from agent.archive import make_archive
+from common.backup_request import BackupRequest
+
 from .config import log, config
-from common.error import SSD_ConfigMissingParam
+from common.error import *
 from .node import Node
 
 DEFAULT_ENTRY_PARAM = {
@@ -21,18 +24,24 @@ DEFAULT_ENTRY_PARAM = {
     "fallback" : None
 }
 
-def get_opt(entry, name):
-    try:
-        return config[entry, name]
-    except:
-        pass
+def get_opt(entry, name=None):
+    if name:
+        try:
+            return config[entry, name]
+        except:
+            pass
 
-    try:
-        return config["global", name]
-    except:
-        pass
+        try:
+            return config["global", name]
+        except:
+            pass
 
-    if DEFAULT_ENTRY_PARAM[name] : return DEFAULT_ENTRY_PARAM[name]
+        if DEFAULT_ENTRY_PARAM[name] : return DEFAULT_ENTRY_PARAM[name]
+    else:
+        try:
+            return config["global", entry]
+        except:
+            pass
 
     raise SSD_ConfigMissingParam("Erreur dans l'entrée '%s' le paramètre '%s' n'est pas défini"%(entry, name))
 
@@ -44,6 +53,7 @@ class Entry:
             "name": get_opt(entry, "name"),
             "path": get_opt(entry, "path"),
             "node": get_opt(entry, "node"),
+            "agent_url" : get_opt("url"),
             "temp_dir": get_opt(entry, "temp_dir"),
             "fallback": get_opt(entry, "fallback")
         })
@@ -51,6 +61,8 @@ class Entry:
     def __init__(self, data : dict):
         self.name = data["name"]
         self.dirs = data["path"]
+        self.agent_url = data["agent_url"]
+        self.current_node = None
         node = data["node"]
         fallback = data["fallback"] if data["fallback"] else []
         self.temp_dir = data["backup_dir"] if "backup_dir" in data else "temp"
@@ -64,34 +76,50 @@ class Entry:
         return path, hash
 
 
+    def find_node_by_url(self, url : str):
+        for node in self.nodes:
+            if node.url==url:
+                return node
+        return None
+
     def get_token(self, path, hash):
+        self.current_node = None
         size = os.path.getsize(path)
         data = {
             "time" : time.time(),
             "agent" : config["global", "site"],
             "from" : "Agent/"+config["global", "site"],
             "size" : size,
-            "hash" : hash
+            "hash" : hash,
+            "backup_name" : self.name,
+            "agent_url" : self.agent_url
         }
 
         log.info("Requête de sauvegarde")
         for node in self.nodes:
             log.debug("Demande à '%s'" % str(node))
-            allowed = node.request_backup(data)
-            if allowed:
-                log.debug("\tSauvegarde autorisée")
-                return True
-            log.debug("\tSauvegarde impossible")
+            ret = node.request_backup(data)
+            if ret.ok():
+                self.current_node = node
+                log.debug("\tSauvegarde autorisée sur '%s'"%node.url)
+                return ret
+            log.debug("\tSauvegarde impossible sur '%s', raison: '%s'"%(node.url, ret.message))
         log.critical("Erreur impossible de faire la sauvegarde, aucun noeud n'a pu répondre à la requête")
-        return False
+        return SSDE_NoNodeAvailable()
 
-    def upload(self, file, res):
-
-
-        return res
+    def upload(self, file :str, token : str):
+        ret = self.current_node.upload(file, token)
+        if ret.err():  return ret
+        self.current_node=None
+        return ret
 
     def backup(self):
         path, hash = self.create_archive()
         res = self.get_token(path, hash)
-        res = self.upload(path, res)
+
+        if not self.current_node or res.err(): return res
+
+        token = res["token"]
+        res = self.upload(path, token)
+
         os.remove(path)
